Move detail view data loaders to utils
