+++ packages/keycloak-user-management/src/components/UserGroupsList/index.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet';
 import { Row, Col, Button, Table, Spin, Divider, Dropdown, Menu, PageHeader } from 'antd';
 import { Link } from 'react-router-dom';
-import { Redirect, RouteComponentProps, useHistory } from 'react-router';
+import { RouteComponentProps, useHistory } from 'react-router';
 import { MoreOutlined, PlusOutlined } from '@ant-design/icons';
 import { useDispatch, useSelector } from 'react-redux';
 import reducerRegistry from '@onaio/redux-reducer-registry';
@@ -36,6 +36,7 @@ import {
   makeKeycloakUserGroupsSelector,
 } from '../../ducks/userGroups';
 import { ViewDetails } from '../UserGroupDetailView';
+import { loadGroupDetails, loadGroupMembers } from './utils';
 
 /** Register reducer */
 reducerRegistry.register(keycloakUserGroupsReducerName, keycloakUserGroupsReducer);
@@ -76,47 +77,6 @@ interface Props {
 /** default component props */
 const defaultProps = {
   keycloakBaseURL: '',
-};
-
-/** Function to fetch group members from keycloak
- *
- * @param {string} groupId - user group id
- * @param {string} baseURL - keycloak base url
- * @param {Function} callback - callback function to set group members from api to state
- */
-export const loadGroupMembers = async (
-  groupId: string,
-  baseURL: string,
-  callback: (members: UserGroupMembers[]) => void
-) => {
-  const serve = new KeycloakService(`${KEYCLOAK_URL_USER_GROUPS}/${groupId}/members`, baseURL);
-  return await serve
-    .list()
-    .then((response: UserGroupMembers[]) => {
-      callback(response);
-    })
-    .catch((e: Error) => sendErrorNotification(`${e}`))
-    .finally(() => <Redirect to={`${URL_USER_GROUPS}/${groupId}`} />);
-};
-
-/** Function to fetch group members from keycloak
- *
- * @param {string} groupId - user group id
- * @param {string} baseURL - keycloak base url
- * @param {Function} callback - callback function to set group members from api to state
- */
-export const loadGroupDetails = async (
-  groupId: string,
-  baseURL: string,
-  callback: (userGroups: KeycloakUserGroup) => void
-) => {
-  const serve = new KeycloakService(KEYCLOAK_URL_USER_GROUPS, baseURL);
-  return await serve
-    .read(groupId)
-    .then((response: KeycloakUserGroup) => {
-      callback(response);
-    })
-    .catch((e: Error) => sendErrorNotification(`${e}`));
 };
 
 export type UserGroupListTypes = Props & RouteComponentProps<RouteParams>;
